# Feature: Append new submission data by chunks to the triple store