@jonquet 2022-07-01
CC: @alexskr  with who I discussed the problem (loading huge file in 4store) last April. 
The proposed solution seems a good practice for groups (like us) running the Appliance and then hosting 4store on the same machine.