@alexskr 2022-09-14
The current non-chunked RDF upload approach appropriately handles situations where triple store fails uploads of generated RDF due to malformed data errors, things as the mismatched types which owlapi doesn't catch (see https://github.com/ncbo/bioportal-project/issues/253)
The whole operation fails so the triple store doesn't end up with a partially loaded graph; however, with chunked data uploads, it would be possible that a chunked upload fails in the middle and would leave an incomplete graph in the triple store.  Do you have any mitigation mechanisms in place for this kind of problem?