@syphax-bouazzouni 2022-09-14
HI @alexskr,

I think if one of the chunks fails,  a `RestClient::BadRequest` will be raised and stop the process (like https://github.com/ncbo/bioportal-project/issues/253).

And when we reprocess it again it will delete the remained graph and create a new empty one for appending again the chunks from the start.